src/requirements.txt
Django==4.1.2
djangorestframework==3.14.0
psycopg2-binary==2.9.4
django3-livesync==1 # DEBUG
django-extensions==3.2.1 # DEBUG
tzdata==2022.4 # FOR ADMIN
